backend/requirements.txt
flask
openai==1.34.0
python-dotenv